Add pre-commit tooling, clang-format, and CI workflows
Reformats main.c to clang-format's Google style and hardens
days_in_zone to use int64_t explicitly. Adds pre-commit hooks
(clang-format, cpplint, codespell, basic hygiene checks) with a
pre-commit CI job, and a publish workflow that builds and ships
the watchface to the Rebble app store on push to main.

+++ src/c/main.c
@@ -28,20 +28,20 @@
 
 // MAX_ZONES is the user-configurable extra zones. Two more rows (local + UTC)
 // are always drawn, so MAX_ROWS bounds every row loop in this file.
-#define MAX_ZONES      6
-#define MAX_ROWS       (MAX_ZONES + 2)
-#define MAX_LABEL_LEN  12
+#define MAX_ZONES 6
+#define MAX_ROWS (MAX_ZONES + 2)
+#define MAX_LABEL_LEN 12
 
 #define PERSIST_LOCAL_OFFSET 1
-#define PERSIST_NUM_ZONES    2
-#define PERSIST_ZONES_BLOB   3
-#define PERSIST_H24          4
-#define PERSIST_DARK         5
-#define PERSIST_LOCAL_LABEL  6
+#define PERSIST_NUM_ZONES 2
+#define PERSIST_ZONES_BLOB 3
+#define PERSIST_H24 4
+#define PERSIST_DARK 5
+#define PERSIST_LOCAL_LABEL 6
 
 typedef struct {
-  char    label[MAX_LABEL_LEN + 1];
-  int32_t offset_min;               // minutes east of UTC
+  char label[MAX_LABEL_LEN + 1];
+  int32_t offset_min;  // minutes east of UTC
 } Zone;
 
 // AppMessage keys for per-slot zone values. A JS-side array fans out to
@@ -68,15 +68,15 @@ static void init_message_keys(void) {
 
 // ---- state ----
 static Window *s_window;
-static Layer  *s_header_layer;
-static Layer  *s_rows_layer;
+static Layer *s_header_layer;
+static Layer *s_rows_layer;
 
-static Zone    s_zones[MAX_ZONES];
-static int     s_num_zones        = 0;
+static Zone s_zones[MAX_ZONES];
+static int s_num_zones = 0;
 static int32_t s_local_offset_min = 0;
-static char    s_local_label[MAX_LABEL_LEN + 1] = "LOCAL";
-static bool    s_h24              = true;
-static bool    s_dark             = false;
+static char s_local_label[MAX_LABEL_LEN + 1] = "LOCAL";
+static bool s_h24 = true;
+static bool s_dark = false;
 
 // -----------------------------------------------------------------------------
 // Theme
@@ -86,9 +86,15 @@ static bool    s_dark             = false;
 // (blue) bands are fixed — they're the point of the watchface, and their white
 // text reads well on both.
 
-static GColor theme_bg(void)  { return s_dark ? GColorBlack : GColorWhite; }
-static GColor theme_fg(void)  { return s_dark ? GColorWhite : GColorBlack; }
-static GColor theme_rule(void) { return s_dark ? GColorDarkGray : GColorLightGray; }
+static GColor theme_bg(void) {
+  return s_dark ? GColorBlack : GColorWhite;
+}
+static GColor theme_fg(void) {
+  return s_dark ? GColorWhite : GColorBlack;
+}
+static GColor theme_rule(void) {
+  return s_dark ? GColorDarkGray : GColorLightGray;
+}
 
 // -----------------------------------------------------------------------------
 // Row fonts
@@ -101,23 +107,24 @@ static GColor theme_rule(void) { return s_dark ? GColorDarkGray : GColorLightGra
 
 typedef struct {
   const char *key;
-  int         line_h;
-  int         min_row_h;   // smallest band this font is allowed in
-  int         time_w_24;
-  int         time_w_12;
+  int line_h;
+  int min_row_h;  // smallest band this font is allowed in
+  int time_w_24;
+  int time_w_12;
 } RowFont;
 
 static const RowFont ROW_FONTS[] = {
-  { FONT_KEY_GOTHIC_28_BOLD, 30, 40, 76, 92 },
-  { FONT_KEY_GOTHIC_24_BOLD, 26, 32, 66, 80 },
-  { FONT_KEY_GOTHIC_18_BOLD, 20, 23, 50, 62 },
-  { FONT_KEY_GOTHIC_14_BOLD, 16,  0, 40, 50 },
+    {FONT_KEY_GOTHIC_28_BOLD, 30, 40, 76, 92},
+    {FONT_KEY_GOTHIC_24_BOLD, 26, 32, 66, 80},
+    {FONT_KEY_GOTHIC_18_BOLD, 20, 23, 50, 62},
+    {FONT_KEY_GOTHIC_14_BOLD, 16, 0, 40, 50},
 };
 #define NUM_ROW_FONTS ((int)(sizeof(ROW_FONTS) / sizeof(ROW_FONTS[0])))
 
 static const RowFont *row_font_for(int row_h) {
   for (int i = 0; i < NUM_ROW_FONTS; i++) {
-    if (row_h >= ROW_FONTS[i].min_row_h) return &ROW_FONTS[i];
+    if (row_h >= ROW_FONTS[i].min_row_h)
+      return &ROW_FONTS[i];
   }
   return &ROW_FONTS[NUM_ROW_FONTS - 1];  // unreachable: last entry has min 0
 }
@@ -142,9 +149,12 @@ static void load_defaults(void) {
   set_label(s_local_label, "LOCAL");
   memset(s_zones, 0, sizeof(s_zones));
   s_num_zones = 3;
-  set_label(s_zones[0].label, "NYC"); s_zones[0].offset_min = -300;
-  set_label(s_zones[1].label, "LDN"); s_zones[1].offset_min = 0;
-  set_label(s_zones[2].label, "TYO"); s_zones[2].offset_min = 540;
+  set_label(s_zones[0].label, "NYC");
+  s_zones[0].offset_min = -300;
+  set_label(s_zones[1].label, "LDN");
+  s_zones[1].offset_min = 0;
+  set_label(s_zones[2].label, "TYO");
+  s_zones[2].offset_min = 540;
 }
 
 static void load_config(void) {
@@ -154,8 +164,10 @@ static void load_config(void) {
   }
 
   s_num_zones = persist_read_int(PERSIST_NUM_ZONES);
-  if (s_num_zones < 0)         s_num_zones = 0;
-  if (s_num_zones > MAX_ZONES) s_num_zones = MAX_ZONES;
+  if (s_num_zones < 0)
+    s_num_zones = 0;
+  if (s_num_zones > MAX_ZONES)
+    s_num_zones = MAX_ZONES;
   persist_read_data(PERSIST_ZONES_BLOB, s_zones, sizeof(s_zones));
 
   if (persist_exists(PERSIST_LOCAL_OFFSET)) {
@@ -181,11 +193,11 @@ static void load_config(void) {
 
 static void save_config(void) {
   persist_write_int(PERSIST_LOCAL_OFFSET, s_local_offset_min);
-  persist_write_int(PERSIST_NUM_ZONES,    s_num_zones);
-  persist_write_int(PERSIST_H24,          s_h24 ? 1 : 0);
-  persist_write_int(PERSIST_DARK,         s_dark ? 1 : 0);
+  persist_write_int(PERSIST_NUM_ZONES, s_num_zones);
+  persist_write_int(PERSIST_H24, s_h24 ? 1 : 0);
+  persist_write_int(PERSIST_DARK, s_dark ? 1 : 0);
   persist_write_string(PERSIST_LOCAL_LABEL, s_local_label);
-  persist_write_data(PERSIST_ZONES_BLOB,  s_zones, sizeof(s_zones));
+  persist_write_data(PERSIST_ZONES_BLOB, s_zones, sizeof(s_zones));
 }
 
 // -----------------------------------------------------------------------------
@@ -196,9 +208,10 @@ static void save_config(void) {
 // differencing two zones, so the absolute value doesn't matter — but it does
 // have to floor towards negative infinity for pre-1970 sanity.
 static int days_in_zone(time_t utc_now, int32_t off_min) {
-  long long shifted = (long long)utc_now + (long long)off_min * 60LL;
-  long long d = shifted / 86400LL;
-  if (shifted < 0 && (shifted % 86400LL) != 0) d--;
+  int64_t shifted = (int64_t)utc_now + (int64_t)off_min * 60LL;
+  int64_t d = shifted / 86400LL;
+  if (shifted < 0 && (shifted % 86400LL) != 0)
+    d--;
   return (int)d;
 }
 
@@ -209,7 +222,8 @@ static void format_zone_time(time_t utc_now, int32_t off_min, char *out, int n) 
     snprintf(out, n, "%02d:%02d", t->tm_hour, t->tm_min);
   } else {
     int h = t->tm_hour % 12;
-    if (h == 0) h = 12;
+    if (h == 0)
+      h = 12;
     snprintf(out, n, "%d:%02d%s", h, t->tm_min, t->tm_hour >= 12 ? "p" : "a");
   }
 }
@@ -223,21 +237,23 @@ static void format_zone_time(time_t utc_now, int32_t off_min, char *out, int n) 
 
 typedef struct {
   const char *label;
-  int32_t     offset_min;
-  GColor      bg;
-  GColor      fg;
+  int32_t offset_min;
+  GColor bg;
+  GColor fg;
 } Row;
 
 static int build_rows(Row *rows, int cap) {
   int n = 0;
 
   if (n < cap) {
-    rows[n++] = (Row){ .label = s_local_label, .offset_min = s_local_offset_min,
-                       .bg = GColorRed, .fg = GColorWhite };
+    rows[n++] = (Row){.label = s_local_label,
+                      .offset_min = s_local_offset_min,
+                      .bg = GColorRed,
+                      .fg = GColorWhite};
   }
   if (n < cap) {
-    rows[n++] = (Row){ .label = "UTC", .offset_min = 0,
-                       .bg = GColorBlue, .fg = GColorWhite };
+    rows[n++] =
+        (Row){.label = "UTC", .offset_min = 0, .bg = GColorBlue, .fg = GColorWhite};
   }
   // A configured zone that currently resolves to the local or UTC offset would
   // just repeat a band that's already on screen — Lisbon in winter is UTC, and
@@ -246,9 +262,10 @@ static int build_rows(Row *rows, int cap) {
   // collide for half the year and separate again when DST shifts.
   for (int i = 0; i < s_num_zones && i < MAX_ZONES && n < cap; i++) {
     int32_t off = s_zones[i].offset_min;
-    if (off == s_local_offset_min || off == 0) continue;
-    rows[n++] = (Row){ .label = s_zones[i].label, .offset_min = off,
-                       .bg = theme_bg(), .fg = theme_fg() };
+    if (off == s_local_offset_min || off == 0)
+      continue;
+    rows[n++] = (Row){
+        .label = s_zones[i].label, .offset_min = off, .bg = theme_bg(), .fg = theme_fg()};
   }
 
   // Sort east-to-west: everything ahead of UTC sits above the UTC band, and
@@ -285,13 +302,13 @@ static void header_layer_update(Layer *layer, GContext *ctx) {
   char date_buf[20];
   strftime(date_buf, sizeof(date_buf), "%a %d %b", t);
 
-  GFont font = fonts_get_system_font(large ? FONT_KEY_GOTHIC_24_BOLD
-                                           : FONT_KEY_GOTHIC_18_BOLD);
+  GFont font =
+      fonts_get_system_font(large ? FONT_KEY_GOTHIC_24_BOLD : FONT_KEY_GOTHIC_18_BOLD);
   int line_h = large ? 26 : 20;
   graphics_context_set_text_color(ctx, theme_fg());
   graphics_draw_text(ctx, date_buf, font,
-      GRect(4, (b.size.h - line_h) / 2 - 3, b.size.w - 8, line_h + 6),
-      GTextOverflowModeTrailingEllipsis, GTextAlignmentCenter, NULL);
+                     GRect(4, (b.size.h - line_h) / 2 - 3, b.size.w - 8, line_h + 6),
+                     GTextOverflowModeTrailingEllipsis, GTextAlignmentCenter, NULL);
 
   // Hairline under the header, separating it from the first band.
   graphics_context_set_stroke_color(ctx, theme_rule());
@@ -303,25 +320,26 @@ static void rows_layer_update(Layer *layer, GContext *ctx) {
 
   Row rows[MAX_ROWS];
   int n = build_rows(rows, MAX_ROWS);
-  if (n <= 0) return;
+  if (n <= 0)
+    return;
 
   // Bands are laid out by interpolating the band edges across the full body
   // height. That spreads the leftover pixels evenly and, more importantly,
   // leaves no uncoloured seams between adjacent bands.
   int body_h = bounds.size.h;
-  int row_h  = body_h / n;
+  int row_h = body_h / n;
 
   const RowFont *rf = row_font_for(row_h);
-  GFont  font   = fonts_get_system_font(rf->key);
-  int    time_w = s_h24 ? rf->time_w_24 : rf->time_w_12;
+  GFont font = fonts_get_system_font(rf->key);
+  int time_w = s_h24 ? rf->time_w_24 : rf->time_w_12;
 
   time_t utc_now = time(NULL);
-  int    local_day = days_in_zone(utc_now, s_local_offset_min);
+  int local_day = days_in_zone(utc_now, s_local_offset_min);
 
   for (int i = 0; i < n; i++) {
     int y0 = (body_h * i) / n;
     int y1 = (body_h * (i + 1)) / n;
-    int h  = y1 - y0;
+    int h = y1 - y0;
 
     graphics_context_set_fill_color(ctx, rows[i].bg);
     graphics_fill_rect(ctx, GRect(0, y0, bounds.size.w, h), 0, GCornerNone);
@@ -351,11 +369,11 @@ static void rows_layer_update(Layer *layer, GContext *ctx) {
 
     graphics_context_set_text_color(ctx, rows[i].fg);
     graphics_draw_text(ctx, label_buf, font,
-        GRect(5, text_y, bounds.size.w - time_w - 8, text_h),
-        GTextOverflowModeTrailingEllipsis, GTextAlignmentLeft, NULL);
+                       GRect(5, text_y, bounds.size.w - time_w - 8, text_h),
+                       GTextOverflowModeTrailingEllipsis, GTextAlignmentLeft, NULL);
     graphics_draw_text(ctx, time_buf, font,
-        GRect(bounds.size.w - time_w - 5, text_y, time_w, text_h),
-        GTextOverflowModeTrailingEllipsis, GTextAlignmentRight, NULL);
+                       GRect(bounds.size.w - time_w - 5, text_y, time_w, text_h),
+                       GTextOverflowModeTrailingEllipsis, GTextAlignmentRight, NULL);
   }
 }
 
@@ -370,7 +388,8 @@ static void redraw_all(void) {
 
 static void request_config_refresh(void) {
   DictionaryIterator *iter;
-  if (app_message_outbox_begin(&iter) != APP_MSG_OK) return;
+  if (app_message_outbox_begin(&iter) != APP_MSG_OK)
+    return;
   dict_write_int32(iter, MESSAGE_KEY_REQUEST_CONFIG, 1);
   app_message_outbox_send();
 }
@@ -386,7 +405,10 @@ static void inbox_received_handler(DictionaryIterator *iter, void *context) {
   bool changed = false;
 
   Tuple *t = dict_find(iter, MESSAGE_KEY_LOCAL_OFFSET);
-  if (t) { s_local_offset_min = t->value->int32; changed = true; }
+  if (t) {
+    s_local_offset_min = t->value->int32;
+    changed = true;
+  }
 
   t = dict_find(iter, MESSAGE_KEY_LOCAL_LABEL);
   if (t && t->type == TUPLE_CSTRING) {
@@ -395,16 +417,24 @@ static void inbox_received_handler(DictionaryIterator *iter, void *context) {
   }
 
   t = dict_find(iter, MESSAGE_KEY_H24);
-  if (t) { s_h24 = (t->value->int32 != 0); changed = true; }
+  if (t) {
+    s_h24 = (t->value->int32 != 0);
+    changed = true;
+  }
 
   t = dict_find(iter, MESSAGE_KEY_DARK);
-  if (t) { s_dark = (t->value->int32 != 0); changed = true; }
+  if (t) {
+    s_dark = (t->value->int32 != 0);
+    changed = true;
+  }
 
   t = dict_find(iter, MESSAGE_KEY_NUM_ZONES);
   if (t) {
     int n = t->value->int32;
-    if (n < 0)         n = 0;
-    if (n > MAX_ZONES) n = MAX_ZONES;
+    if (n < 0)
+      n = 0;
+    if (n > MAX_ZONES)
+      n = MAX_ZONES;
     s_num_zones = n;
     changed = true;
   }
@@ -422,10 +452,11 @@ static void inbox_received_handler(DictionaryIterator *iter, void *context) {
     }
   }
 
-  if (!changed) return;
+  if (!changed)
+    return;
 
-  APP_LOG(APP_LOG_LEVEL_INFO, "config updated: zones=%d local_off=%ld h24=%d dark=%d",
-          s_num_zones, (long)s_local_offset_min, (int)s_h24, (int)s_dark);
+  APP_LOG(APP_LOG_LEVEL_INFO, "config updated: zones=%d local_off=%d h24=%d dark=%d",
+          s_num_zones, (int)s_local_offset_min, (int)s_h24, (int)s_dark);
   save_config();
   window_set_background_color(s_window, theme_bg());
   redraw_all();
@@ -436,9 +467,9 @@ static void inbox_received_handler(DictionaryIterator *iter, void *context) {
 // -----------------------------------------------------------------------------
 
 static void window_load(Window *window) {
-  Layer *root  = window_get_root_layer(window);
+  Layer *root = window_get_root_layer(window);
   GRect bounds = layer_get_bounds(root);
-  int   hdr_h  = header_height(bounds.size.h);
+  int hdr_h = header_height(bounds.size.h);
 
   window_set_background_color(window, theme_bg());
 
@@ -462,10 +493,10 @@ static void init(void) {
   load_config();
 
   s_window = window_create();
-  window_set_window_handlers(s_window, (WindowHandlers) {
-    .load   = window_load,
-    .unload = window_unload,
-  });
+  window_set_window_handlers(s_window, (WindowHandlers){
+                                           .load = window_load,
+                                           .unload = window_unload,
+                                       });
   window_stack_push(s_window, true);
 
   tick_timer_service_subscribe(MINUTE_UNIT, tick_handler);
